chore(language/test): update to new diagnostic model

+++ language/test/src/query/runner/navigation/call_hierarchy.rs
@@ -432,7 +432,7 @@ fn validate_call_ranges(
     }
 
     // validate ordering and duplicates within the call site ranges
-    let mut previous: Option<(u64, u32, u32)> = None;
+    let mut previous: Option<(u128, u32, u32)> = None;
     for span in ranges {
         // build a stable ordering key for the call site span
         let key = span_key(*span);
@@ -459,7 +459,7 @@ fn validate_call_ordering<'a>(
     items: impl Iterator<Item = &'a CallHierarchyItem>,
     errors: &mut Vec<String>,
 ) {
-    let mut previous: Option<(u64, u32, u32, u32, u32, u8, String)> = None;
+    let mut previous: Option<(u128, u32, u32, u32, u32, u8, String)> = None;
 
     // compare each item key against the previous one
     for item in items {
@@ -508,7 +508,7 @@ fn validate_span_bounds(
 }
 
 /// Build a stable ordering key for a call hierarchy item.
-fn item_key(item: &CallHierarchyItem) -> (u64, u32, u32, u32, u32, u8, String) {
+fn item_key(item: &CallHierarchyItem) -> (u128, u32, u32, u32, u32, u8, String) {
     (
         item.file.0,
         item.range.start,
@@ -521,6 +521,6 @@ fn item_key(item: &CallHierarchyItem) -> (u64, u32, u32, u32, u32, u8, String) {
 }
 
 /// Build a stable ordering key for a span.
-fn span_key(span: Span) -> (u64, u32, u32) {
+fn span_key(span: Span) -> (u128, u32, u32) {
     (span.file.0, span.start, span.end)
 }
